requirements.txt
streamlit
pyautogen
python-dotenv
pandas
tabulate
streamlit-authenticator
azure-identity
azure-search-documents